# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI营销战略执行智能体",
    "description": "用AI帮小企业把营销战略翻译成每天的执行动作",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/plan/{plan_id}/confirm": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Confirm Plan",
        "description": "确认计划 → 标记 confirmed 并自动排期落库 todos。",
        "operationId": "confirm_plan_api_plan__plan_id__confirm_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plan/{plan_id}/edit": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Edit Plan",
        "description": "微调计划中的任务（标题/时段/做法/清单），写回计划的 days JSON。",
        "operationId": "edit_plan_api_plan__plan_id__edit_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit Plan Api Plan  Plan Id  Edit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plan/{plan_id}/regenerate": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Regenerate Plan",
        "description": "基于同一企业重新生成计划（结合已有记忆/反馈）。",
        "operationId": "regenerate_plan_api_plan__plan_id__regenerate_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Regenerate Plan Api Plan  Plan Id  Regenerate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedule/checkin": {
      "put": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Checkin Todo",
        "description": "任务完成打卡 / 修改状态，落库 todos（闭环：完成标记反馈至复盘 Agent）。",
        "operationId": "checkin_todo_api_schedule_checkin_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Checkin Todo Api Schedule Checkin Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedule/sync": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Sync Schedule",
        "description": "把排期结果同步落库（用于「安排本周日程」快捷指令或确认后补录）。",
        "operationId": "sync_schedule_api_schedule_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Sync Schedule Api Schedule Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedule": {
      "get": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Get Schedule",
        "description": "读取当前企业已落库的排期待办（跨会话持久，反映打卡状态）。",
        "operationId": "get_schedule_api_schedule_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Schedule Api Schedule Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/upload": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Files Upload",
        "description": "通用文件暂存：小程序本地临时路径 → 服务端可读路径。\n\n返回的 ``file_path`` 可直接放进 ``/agent/chat`` 或 ``/agent/chat/stream``\n的 ``files`` 字段——这是「对话里发数据 → Agent 真能读到」的关键一步，\n否则小程序传过来的只是设备本地临时路径，服务端根本打不开。",
        "operationId": "files_upload_api_files_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_files_upload_api_files_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Files Upload Api Files Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/review/trigger": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Review Trigger",
        "description": "生成本周复盘报告 + 下周建议，并写入 ReviewRecord。\n\n优先使用用户上传的指标作为实际值；若无上传数据则提示先上传。",
        "operationId": "review_trigger_api_review_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Review Trigger Api Review Trigger Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/{review_id}/apply": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Review Apply",
        "description": "把复盘建议直接应用到下周：重新生成计划并自动排期。",
        "operationId": "review_apply_api_review__review_id__apply_post",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Apply Api Review  Review Id  Apply Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/feedback": {
      "post": {
        "tags": [
          "闭环业务"
        ],
        "summary": "Agent Feedback",
        "description": "采集用户对 Agent 产出的反馈（赞/踩/计划修改），更新策略有效性评分。",
        "operationId": "agent_feedback_api_agent_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Agent Feedback Api Agent Feedback Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_files_upload_api_files_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_files_upload_api_files_upload_post"
      },
...
      "CheckinRequest": {
        "properties": {
          "todo_id": {
            "type": "string",
            "title": "Todo Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "todo_id"
        ],
        "title": "CheckinRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanConfirmResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id",
            "default": ""
          },
          "schedule": {
            "items": {},
            "type": "array",
            "title": "Schedule",
            "default": []
          }
        },
        "type": "object",
        "title": "PlanConfirmResponse"
      },
...
      "PlanEditRequest": {
        "properties": {
          "edits": {
            "items": {
              "$ref": "#/components/schemas/PlanEditItem"
            },
            "type": "array",
            "title": "Edits",
            "default": []
          }
        },
        "type": "object",
        "title": "PlanEditRequest"
      },
...
      "ReviewApplyRequest": {
        "properties": {
          "business_id": {
            "type": "string",
            "title": "Business Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ReviewApplyRequest"
      },
...
      "ReviewTriggerRequest": {
        "properties": {
          "business_id": {
            "type": "string",
            "title": "Business Id",
            "default": ""
          },
          "week_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Number"
          }
        },
        "type": "object",
        "title": "ReviewTriggerRequest"
      },
...
      "ScheduleSyncRequest": {
        "properties": {
          "business_id": {
            "type": "string",
            "title": "Business Id",
            "default": ""
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "days": {
            "items": {},
            "type": "array",
            "title": "Days",
            "default": []
          }
        },
        "type": "object",
        "title": "ScheduleSyncRequest"
      },
...
    }
  }
}